Support FASTA headers without whitespace
Allow support for indexing FASTA files with identifiers that don't
contain whitespace while `full_name=False`.

Index creation iterates through the header line until a whitespace
character is found, but if the iteration doesn't encounter a whitespace
character, the string is terminated at a position one-too-far. If this
happens, just reset the index back to the length of the string.